+++ src/objects/impls/mineral.rs
@@ -5,8 +5,7 @@ use crate::{
 
 simple_accessors! {
     impl Mineral {
-        pub fn density() -> u32 = density;
-        pub fn mineral_amount() -> Density = mineralAmount;
+        pub fn density() -> Density = density;
         // id from HasId trait
         pub fn ticks_to_regeneration() -> u32 = ticksToRegeneration;
     }
@@ -15,5 +14,10 @@ simple_accessors! {
 impl Mineral {
     pub fn mineral_type(&self) -> ResourceType {
         js_unwrap!(__resource_type_str_to_num(@{self.as_ref()}.mineralType))
+    }
+
+    pub fn mineral_amount(&self) -> u32 {
+        // workaround for the fact that some private servers return floating point mineralAmount values
+        js_unwrap!(Math.floor(@{self.as_ref()}.mineralAmount))
     }
 }