+++ CHANGELOG.md
@@ -43,6 +43,9 @@ Unreleased
 - Fix `Room::find_path` function call to underlying javascript
 - Fix typo in `Position::create_named_construction_site` and work around screeps bug in
   `Room::create_named_construction_site` by passing x and y instead of position object
+- Fix javascript associated object name for `StructureSpawn::spawning`
+- Correct swapped return types for `Mineral::density()` and `Mineral::mineral_amount()` and add
+  a workaround for some private servers returning floating point `mineralAmount` values
 - Fix typo in `StructureController::reservation()` ticks_to_end return value
 
 0.7.0 (2019-10-19)